Complete multi lang support & sample file

+++ sampleFile/cpp.cpp
@@ -13,7 +13,7 @@ int main() {
 
   for(int i=0; i<N; i++) hello.push_back("Hello");
 
-  for(int i=0; i<hello.size(); i++) cout << hello[i] << endl;
+  for(int i=0; i<(int) hello.size(); i++) cout << hello[i] << endl;
 }
 
 /*
